Separate JiraIssues and Products classes from Database class. Rename to JiraIssuesEntity and ProductsEntity
